[internal_lib] Support parsing shelves on systems with no Python2
On macOS & some Linux distributions `python` is legacy Python 2, and
`python3` is Python 3. However, on other Linux distributions `python` is
Python 3 and there is no Python 2.

One of the changes from Python 2 to 3 was removing `execfile()` in favor of
`exec()`. This commit uses the correct one depending on Python version.