Add free-form attributes (e.g., cacheability, shareability, etc) to the PTE (#129)
* [lib] Add support for memory attributes in PTEVal

Add a new field attrs in the PTEVal struct. This field a free-form
list of strings and the specific values will depend on the
architecture. In addition, this change adds support for attributes in
the litmus tests parser. The PTE attributes are parsed as coma
separated strings in the pte.

For example:

pte_x=(oa:phy_y,attrs:(Normal,Inner-write-back,Outer-write-back,Inner-shareable));

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>

* [litmus] Add support for setting PTE attributes in KVM tests

PTE attributes are encoded in a string list in PTEVal; litmus adds the
necessary glue code to set these attributes to the actual PTE via a
helper function define in kvm-headers.h.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>

* [herd] Annotate implicit PTE reads with their memory attributes

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>

* [litmus] Fix AArch64 cacheability attributes naming

AArch64 distinguishes between cacheability attributes for caches in
the inner and outer shareability domains. This patch fixes the
cacheability attributes names to allow that.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>