[herd] KVM, AArch64, insert branching events after PTE read. (#156)
Optionally add "Branching" events in-between page table reads and physical memory accesses or faults. The generation of such events is under control of the `-variant PteBranch` variant.

* [herd] Add -variant PteBranch variant.

This variant will command insertion of "Branching events"
between page table read and physical accesses or faults,
with appriate intra instruction dependencies.

* [herd] KVM, AArch64, insert branching events after PTE read.

Handled instructions are load,stores and amo's.
The general principal for PTE read `a` and memory access `b` is:
  1. Produce a new `branching` event `c`.
  2. Add intra instruction dependances:
     (a) a -data-> c -ctrl-> b
     (b) a -data-> b

Store condition (STXR) dependances are to be revised,
failing CAS is not handled yet.