[herd] Typed print (#171)
Use types for printing values in outcomes. This applies mostly to signed integers, which have to be sign-extended to machine size, and to locations of array type, which are now printed as arrays (`litmus` in standard mode does not yet print arrays, this will be fixed in a forthcoming PR).

* [herd] Type directed printing of final environments (outcomes), integer types.
The only technicality is performing sign-extension for integer types
smaller than machine size.

* [herd] Printing outcomes, mask unsigned integers.
Before unsigned integers were sign extended, which is wrong.

* [herd] Display arrays in outcomes.

* [tests] Add tests for herd typed output.
One test for integer types (L011), another for arrays (L012).