[herd] Better analysis of masking failures.
Completes 9ef5ecca14fc7e8a9c520058ba3fa8db8de27c8a.

First consider that executing LDR W0,[X1] when [X1] holds
an address, result in a failure, as herd7 cannot mask
a symbolic address to 32bits. The failure may originate
from an user error (writing W0 in place of X0), or
from herd7 solver attempt to match read and writes.
Namely, some address may not be determined ---
for instance when address results from a memory read.
As herd7 matches such accesses with every accesses
of opposed polarity, some masking failures may occur,
to be understood as a failed matching attempt
and not as an user error.

To account for the distinction, masking failure is ignored only
when at least one load or store address is still undetermined.

The above test is not bulletproof, as there still
can be real user errors, independently of the presence
of unresolved addressses.