Revert "[herd] Discard executions with wrong rf matching." and better solution. (#188)
This PR implements what I think to be a good solution to the problem of herd spurious crashes.

Consider test T1:
```
AArch64 T1
{
  int *p = &x;
  int x = 1;
  0:X2=p;
}
  P0        ;
LDR X1,[X2] ;
LDR W0,[X1] ;
locations [0:X0;]
```
As the address of the second load is not known without reading memory, herd solver attempts  to match the corresponding read event with all writes. In particular il attempt the match with the initial write to p. This choice leads to a contradiction, as the first load (whose location is known) reads `x` and that `x != p`. However the previous solver attempted masking `x` before discovering the contradiction... As a result herd crashed while the execution should be discarded siliently.

Here is an example where herd7 should arguably crash.
```
AArch64 T2
{
  int *p = &x;
  int x = 1;
  0:X2=p;
  1:X2=p;
}
  P0        | P1          ;
LDR X1,[X2] | STR X2,[X2] ;
LDR W0,[X1] |             ;
locations [0:X0;]
```
Consider matching P1 store with P0 first load. Then, P0 second load legally  reads p as a value, and the masking to 32 bits performed fails. The resulting equations are not contradictory and herd is right to crash, as an indication of a severe problem in test.

* Revert "[herd] Discard executions with wrong rf matching. (#184)"

This reverts commit 9ef5ecca14fc7e8a9c520058ba3fa8db8de27c8a.

* [herd] Delay operation failure in solver.

* [herd] More information displayed with option `-debug solver`

When instructed to do so with `-debug solver -gv`, herd7
additionally displays the candidates discarded because
of contradictory rf choices.

* [tests] Add tests.

Tests mostly check the interaction of indirections and
mask32 failure.