[herd] Simple final write choice for arrays in mixed-size.
Resort to simple final write enumeration when some array
or array element is observed.

The simple mode operates on sub-event only and treat
them as independant; while the optimised mode orders
overlapping 'big' events. In effect, it looks
difficult to consider the initial write of the array
as such an event.