[litmus] Klitmus, memory type oriented  memory allocation.
When global locations memory types are given in test, target C code
will use specufic allocators and de-allocators.

Available for X86_64 only.

Notice: the kernel primitives that set the memory type
for virual pages are absent from the linux-header package.
However, they are exported by the kernel and are available once
the adequate prototypes are provided.