[build] `make check-deps` no-longer requires opam
The previous check-deps implementation was created to check that
`stdint` was installed, and assumed the existance of `opam`.

However, we no-longer require `stdint`, and `opam` is not required to be
installed to build herdtools, so now `make check-deps` just checks for
the presence of `ocaml`, `menhir`, and either `dune` or `ocamlbuild` &
`ocamlfind`.

Tested by applying this patch on a machine that has all the required
dependencies, but NOT `opam`.