[herd] Fix iico-data dependancy in CAS equal case
In CAS, the `rs` register which provide the compared value
receives the value read from memory.
In the equal case, where the value read from memory and the one
in the `rs` register are equal, the iico-data dependancy that
provided data to the write in the `rs` register was comming from
the `rs` register, instead of coming from the value read in memory.