[herd] New "bind" monad operator.
The new operator is a slight variation on the classical
(>>=) data composition operator.

Operators differ in the targets of the intruduced iico_data (entry
point to the second monad event structure). For (>>=), those are the
minimal points of iico_data; for the new operator those are the
minimal point of complete iico.