[herd] KVM, change arrangement of intra-instruction AF-update.
The required arrangement is as follows:
```

                   +----------data---------------->|
                   |                               | Address port(s)
           +- a:Rpte_x -data-> b:Branching --ctrl->|
      data |                       |
   +-------+->  Branching <--ctrl--+
   |               |
  \|/              |
WptexAF:1  <-ctrl- +
```

Noticeable points of the arrangement above are
the presence of two branching events and the
placement of outputs at the top (on events a and b)

The first branching event represents the validity
check of the memory action and controls both
the action and the AF update. The second branching event
represents the check of AF:0 and controls the AF update.

This new arrangement helped clarify monad operators,
and prompted introduction of a second output
field in event structures. The new field, ctrl_output
encodes explicit control output when default
(maximal of iicos) is no longer sufficient.

In the scheme of AF update above, data output is read event a,
while control output is branching event b.