[herd] Double branch (#240)
* [herd] Slight code refactoring of AArch64Sem.

In KVM mode branching event is inserted beetween page
table read and memory action in the "address" chain.

The refactoring consists in having this inserted event
generated in one place only, whatever the memory action is.

* [herd] KVM, change arrangement of intra-instruction AF-update.

The required arrangement is as follows:
```

                   +----------data---------------->|
                   |                               | Address port(s)
           +- a:Rpte_x -data-> b:Branching --ctrl->|
      data |                       |
   +-------+->  Branching <--ctrl--+
   |               |
  \|/              |
WptexAF:1  <-ctrl- +
```

Noticeable points of the arrangement above are
the presence of two branching events and the
placement of outputs at the top (on events a and b)

The first branching event represents the validity
check of the memory action and controls both
the action and the AF update. The second branching event
represents the check of AF:0 and controls the AF update.

This new arrangement helped clarify monad operators,
and prompted introduction of a second output
field in event structures. The new field, ctrl_output
encodes explicit control output when default
(maximal of iicos) is no longer sufficient.

In the scheme of AF update above, data output is read event a,
while control output is branching event b.

* [herd] Clarifying code, replacing `List.append` by infix `@`.

* [herd] Add optional text to branching events.

The text appears in diagrams, it may
help to clarify intricate situations.

* [herd] Spurious events cannot the target of some iico edges.

At the moment spurious events are removed from the input
of the bind_ctrldata_first_outputs, which proves sufficient
for avoiding untimely edges to them.

* [herd] Another organisation of spurious AF updates.

Old mode:
  Memory acessses that read AF:0 always udate AF internally.

New mode (statically enabled by variable `experimental`)
  Memory accesses that read AF:0 update AF internally or
  fire a spurious update.

Moreover the spontaneous generation of spurious updates
is under control of `-variant PhantomOnLoad`.

By default, for a given `<id>`, spurious updates are fired (or not)
  - For pte_`<id>` whose af is initially zero.
  - for every explict write to pte_`<id>`

If activated by variant, for a given `<id>`,
spurious updates are fired (or not)
  - If pte_`<id>` is read in final condition
  - For every explicit read of pte_`<id>`.

* [herd] Fix dependencies of DB intra-instruction update.

DB update is sequenced after AF firing
and should be iico-ctrl dependent on
the branching valid:1 event, while preserving
both outputs of the fist monad.

Hence the usage of a new sequencing
combinator `bind_ctrl_first_outputs`.

* [herd] Fix bug triggered by environment optimisation.

Computation of arch-dependent operations (such as left-shifting
by 12 positions to yield TLB key from address) raise exception
`V.Cst.Result`. This should be caught by anticipated computations.

* [herd] New 'iico' value for the `-optace <tag>` command line setting.

The new setting applies the very effective and a priori
always valid optimisation that consists in _not_ generating
rf relations that contradict the iico relations.