inline variant flag propagated when specified
In commit f78fb880 support was added for specifying variants inline in
litmus tests with the syntax `Variant=foo`.

This commit ensures that this variant information is propagated when
parsing in Herd, so we can use this information during simulation.

A mixed-mode unit test is provided which has `variant=mixed` specified
so we know that this feature works when the user specified this flag.