[herd] Add brackets around global locations in output.
Brackets are not present:
  - For right values
  - In node labels (which also retain old style for special identifiers).

Note: most of the work is done by changing the default of option
`-brackets <bool>` to `true`. As a consequence, brackets are suppressed
by the command-line option `-brackets false`.

Note: locations of array type and array cell references are printed
without brackets. Thats is `t[i]` and not `[t][i]`.