[tools] Handle empty lines while lexing outcomes in logs.
At the moment empty lines in list of outcomes present in logs
may occcur in two situations.

1. Some recent logs have one empty line as list of outcomes.
   This occurs when no location is observed and that the test terminates.
   The situation is not to be confused with empty outcome, which occurs
   when test crashes.

2. Some old logs  have one "cosmetic" empty line at the end of the
   list of outcomes, before the Ok/No validation summary.

In case 1. the empty line should be interpreted as a valid final state,
while in case 2 it should be ignored.

Distinction is made at lexing time by ignoring empty lines that do
not occur first.