[all] Arch-dependent page table entries
[herd] Herd compilation is successful

[litmus] Tools litmus and klitmus compiles successfully

[tools] All tools in `tools` compile successfully

Note: module ParsedPteVal is re-written, in order
to facilitate normalisation of generic
page table entries.

[herd] Abstract arch-dependent operations from op.ml

In practice specific operations are replaced
by an unary constructore `ArchOp`, whose argument
is an arch-specific operation. For now identified
such operations are on page table entries.

As a result the page table entry structure is
abstracted better.

[all] fix PteVal regressions

1. In AArch64PteVal printing of attributes.
2. In pteval parsing, (add output address from location
   in declaration `loc = (...)` only when `...`
   does not set the `oa` field.

[gen] Generators compile without error.