[herd] Extended pco computation.
The pco (co precursor) is a partial order
that the co extends. With command line option
`-optace true` the `pco` relation is computed
from `po & loc`, directly for writes and by
following the inverse of rf for reads.
Assuming SC per location, candidates with a
cyclic pco are rejected early.

This commit extends pco computation by
considering that RMW events are both
reads and writes.