[herd] Initial support for ifetch (#303)
This PR introduced _initial_ support, to be completed:
   1. Fix the problem of instructions that are labelled more than once
   2. And, more important, handle specific, `-variant self` model.

* [herd] Initial support for ifetch in herd (enabled via -variant self).

Supports storing a special value NOP by the address of a label. Identified "overwritable" instructions as Bs or NOPs that have a label. Creates for such instructions initial overwrite events. Supports creating reads-from and from-read edges between overwrites and instructions (achieved by embedding read events into semantics of B).P

* a minor fix to the type of instruction literals

Instruction literals such as B used to take a pair of a processor and a string as an argument, while herd only uses a string as an argument. Removed the proc from literals.

* Addressing Luc's comments

* Addressing Luc's comments; moving herd-specific functionality from lib

* [herd] Handle masking of structure constants as an arch-dependent operation

* added basic tests for -variant self; changed FIXMEs into error messages where appropriate

Co-authored-by: Luc Maranget <Luc.Maranget@inria.fr>