[herd] Normalize memory access to code.
As regards read and write events from and to code instructions,
locations must be equal for a given position in code.
Alas! A given instruction can be labeled several times.

As a solution, accesses whose locations are code labels
are _normalized_ by designating one of the labels of
any instruction to be _its_ label (_i.e_ labels sets
have a canonical element).

The solution has a significant impact on code base:
  1. The normalizing function is passed to `build_semantics`
     functions, so as to normalize the argument of AArch64
     `adr` instruction.
  2. The values bound in the init_state field of test
     have to be normalized too.
  3. Opportunity is seized to add overwritable instruction
     initialisation to init_state field. As a consequence
     those code location have their initial writes produced in
     the normal way and the code in eventsMonad.ml is simplified.

Test A004 reference output is changed to being now correct.