[herd]: Print Timeout for test with -debug timeout
Suppose you run `herd7 -timeout 5.0 expensive_test.litmus` and it times
out on `expensive_test`. Herd silently finishes with `expensive_test` timing out and it is
desirable to be able to have a list of timed-out tests in automated
testing. This patch prints a timeout message to standard error so we can
collect lists of tests which did not finish:

```
> herd7 -timeout 5.0 -debug timeout expensive_test.litmus
...
Timeout for expensive_test.litmus
```