[herd] Preserve instruction annotations in generated FAULT events
In some cases, we need to know if the instruction that generated a
FAULT event had any annotations (for example, was a release
instruction). For this reason, this change adds the annotation
information to the Fault action. Subsequently, it makes sure that herd
considers Fault actions to compute eventsets for architecturally
defined annotations. For good measure, and since we're changing the
Fault action, this change also adds the direction (e.g., read or
write) to it. However, for now we don't do anything with it.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>