[herd] Add support for custom fault handlers
This change adds support for custom fault handlers in herd7. Each core
can have each on fault handler. The fault handler will execute right
after a fault is raised and at this end of its execution it will
exit. For now it behaves as if we were executing in Fatal mode.

In the absence of a custom fault handler, simulation of faults remains
unchanged.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>