[herd] Wider overriding by test configuration.
Configuration parameters (including variants) are set by command-line
options and configuration file settings. Some parameters
can be re-defined in test source and take precedence.

This PR limit the scope of initial configuration parameters, so as
to ensure overriding. The change may be useful in the future when
parameters other than variants are overridden in test source