[herd] Add different types of fault and the corresponding cat sets
This PR adds support for a new module, FaultType where the different
types of faults can be defined per architecture. In AArch64, it
defines Data Aborts and Tag Check faults. Then, it adds this as a
field to the fault type and the Fault action and uses that information
to build the corresponding event sets in cat. In AArch64 the two added
sets are EXC-DATA-ABORT and EXC-TAG-CHECK.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>