[herd-www] Reset variant for each test.
It seems that there is only one instance of the Opts module
for all runs of jherd. As a result variants cumulate from
run to run. As a fix, variant is reset before reading config file.
This is a quick fix, not a final solution.

We also follow the simplification of model invocation performed
in herd (cf PR #455)