[herd] Add cat support for logical operations on variants
This change adds support for the logical operators '&&', '||' and
'not' in statements where the condition was previously a variant. This
way, where previously we could use clauses of the form:

if VARIANT then ... else ...

now, we can use clauses of the form:

if VARIANT_COND then ... else ...

where VARIANT_COND is a VARIANT or a logical operation on multiple
VARIANTs using the defined operators.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>