[AArch64] Initial formalisation of the VMSA - Deprecating ETS and introducing new feature ETS2 - Internal visibility rework - MTE relaxation
This release presents a number of changes, listed below. Those changes should
be reflected in the next issue of the Arm ARM. For more details please
contact memory-model@arm.com.

1. Initial formalisation of the VMSA

This version of the model introduces a number of concepts required to describe
orderings due to accessing page tables. Most notable are the notions of
TTD-read-ordered-before, TLBI-ordered-before, and TLBI-Coherence-after.

This version of the model also handles the notions of Access Flag and Dirty
Bit, and the Hardware Management thereof.

2. Deprecating ETS and Introducing new feature ETS2

As part of the work leading to this initial formalisation of the VMSA, it
transpired that the ETS feature was not giving enough guarantee to be useful to
software. As a consequence, Arm deprecates ETS and introduces a new feature
called ETS2 instead.

This version of the model accounts for those changes.

3. Internal visibility rework

As part of the work leading to this initial formalisation of the VMSA, Arm
decided to rework the way the axioms of the model are presented. More
precisely, most behaviours which were caught by Internal visibility are now
caught by External visibility (including coRR), except for a couple of sanity
checks on Reads-from (viz, coRW1-Exp) and Coherence-after (viz, coW{W,R}-Exp).

4. MTE relaxation

Adjacent to that Internal visibility rework is an MTE relaxation: only RW and
WR pairs are now preserved, and this is reflected in hw-reqs and coWR-MTE
respectively.