[herd] Refactor code indexing
Code is now addressed by integers. That way we
can have pointer arithmetics on code pointers.

First usage is to internalise the computation of
return addresses for subroutine calls. For instance,
The AArch64Sem module now completely executes the `BL` instruction,
while some of the execution was performed by the module
`Mem` before.