[herd] Introduce and use the new `ToInteger` operator
`ToInteger(v1,v2)` yields v1 when v1 is an integer constant
or v2 when when v1 is a non-integer constant and v2 is
an integer constant. Notice that, when v1 is non-integer
the operator fails when v2 is not an integer either.

This operator is used for the computation of flags not to
fail wen comparison is applied to non-integer constants
such as symbols or page table entries. Of course the N,C and V
flags may not be correct w.r.t. hardware in those cases.
However, the Z flag is correct.