[herd,litmus,tool] Implement parsing and printing of instruction constants
Format is : `instr:"`_instruction_`"`. That is, a string prefixed
by the keyword `instr:`.

By exception, the keyword `NOP` is also recognised as the
instruction nop.

As regards litmus, opcodes in output are printed as instructions when
match the opcodes of instructions :
  - given as initial values, or
  - present in final conditions, or
  - pointed by a label that is an argument of some ADR instruction.

Other opcodes are printed as `???`. By exception zero is printed
as `instr:"UDF"`.

Instructions used to overwrite others are no more limited to NOP.

However, for running time to remain decent, such an instruction
 1. Is the initial value of some location, or is pointed to by
    some label given as argument to the ADR instruction,
 2. And, if a branch, have a displacement argument.