[tools] Better management of arrays
Arrays were previously handled as strings in parsing, printing
and comparison, resulting in printing inconsistencies and wrong
comparisons in the case of arrays of integers.

Now, arrays are lexed as lists of their elements and integer elements
are normalised and compared as integers.

The changes result in consistent output (all integers shown in hexadecimal
when specified) and correct equality of arrays.