[AArch64, herd] Support for DIC and IDC features (#490)
* [AArch64] Support for DIC and IDC features

This PR sketches out support for enabling ifetch-related functionality that is indicated in the DIC and IDC bits of Cache-Type Register (CTR_EL0). These features are supposed to grant extra powers to stores: when IDC is enabled, the store does not need to be superseded by DC, and when IDC and DIC are enabled, it does not need to be superseded by DC and IC.

The objective is to be able to indicate in the syntax of litmus test that the features are enabled similarly to TTHM for the virtual memory. That is, the PR adds an option "CacheType" that can have entries "DIC" and "IDC" to enable the functiinality globally, and "Pn:DIC" and "Pn:IDC" to enable it for the `Pn`. In the current solution, enabling DIC automatically enables IDC for the specified scope.

Another objective is to be able to export to cat files the sets `DIC` (and `IDC`) that are meant to represent the subset of `W` such that DIC (IDC) is enabled in the corresponding `Pn`.

The following litmus test illustrates the use of the new option.
```
AArch64 MP.FF+dmb.st+po
CacheType=P0:DIC P0:IDC
{
0:X0=NOP; 0:X1=P1:Lself00; 0:X2=P1:Lself01;
1:X1=P1:Lself00;
}
 P0          | P1         ;
 STR W0,[X1] | Lself01:   ;
 DMB ST      |  B Lself00 ;
 STR W0,[X2] |  MOV W10,#1;
             | Lself00:   ;
             |  B L01     ;
             |  MOV W11,#1;
             | L01:       ;
exists (1:X10=1 /\ 1:X11=0)
```