[herd] Model for several memory accesses
Instructions that perform memory copy impose some
order on read and writes to the same location. Account
for this in model with anew condition `order-iico-memory`,
which may be too strict as it partly overlaps with
other conditions.

Surprisingly, this model strengthening  impacts the diycross/MTE
regression test that operates on a model relaxed enough so that all
litmus tests are allowed. As a consequence, the configuration
file of this test relaxes the new  condition.