[herd,litmus] Introduce NZCV argument to the MRS and MSR instructions
Notice that both the name "NZCV" and the semantics of the `MRS Xt,NZCV`
and `MSR NZCV,Xt` instructions suggest that bit "N" is the high order bit,
bit "Z" is the next bit etc. This entails changing the previous
encoding, where "N" was the low order bit of an ad-hoc register that
represents the N, Z, C and B bits of PSTATE.

As a tribute to how code was written, changes are mininal.