[all]: Add region (and scope) to C tests.
This PR adds region and scope information to C tests.
To than aim. the `extra_data` field of tests structures is
now a list, as C tests already had some extra data of their own.

We also seize opportunity to fix `mlock`, so that it
no more outputs empty files.