[gen] New `default` mode for diy
This new mode is a relaxation of the sc mode. It does not
apply the po_safe optimisation. Basically the po_safe
optimisation rejects some candidates composition when their
W/R extremities and s/d status match a safe "po"
candidate.

For instance, in sc mode, adding PodWR as safe will reject some
of the sequences of two candidates that starts with a write
ends with a read and whose entry and exit locations differ,
_e.g._ `Rfi; DpAddrR`. By contrast the default mode will accpet this
sequence.

The sc mode is also simplified, supressing the little-used
and cryptic ``fence_safe'' feature.