[asl+herd] AArch64+ASL now uses code from the ARM
[asl] Fix identifier declaring with implicit args

[asl] Change function semantics to work on monads

Now functions, and primitives, are interpreted as objects that work on
monads. That is, they take a list of monads as input and return a list
of monads corresponding to the return values.
The final type corresponds to primitives new type:

   type primitive = value m list -> value m list m

[asl] fix unknown interpretation of named types

[asllib] Fix asllib warning handling and fix new

[aarch64+asl] Add new variant ASL+AArch64

This new variant includes automatically generated pseudocode from the
Arm ARM to support .opn like instructions as a litmus test.
That means that we now support instructions from the ARM Arm as litmus
tests.

[asl] rework control dependencies and conditions

[asl] Remove data/addr differenciation

[ASL-herd] Handle correctly DATA ports

[ALS+herd] Change seq combi to fix function calls

[ASL+herd] Rework data dependencies in primitives

[ASL+herd] Rework aslseq to manage ctrl deps

[ASL+herd] Add pseudo-arch && improve ctrl deps

[asl+herd] Improve error message on symbolic

@maranget I now understand your double-load problem correctly:
The issue comes from the second load, where the address is a symbolic
value, and the ASL code put it in a record named `AccDesc` which
contains the address. Records are stored (currently) as herd's
`Constant.ConcreteVector`, which cannot contain any symbolic value.
So we are trying to put the address (i.e. a symbolic value), into a
constant vector, hence the error.

I think this is not something easily fixable (would involve me rewriting
the `ASLValue` module not to use `Constant` nor `symbValue`), so I
suggest keeping this improved error message.

Furthermore, I've added a new ASL regression test with the same
behaviour but in ASL, so without the record problem. It passes without
any problems.

Bug reference:
https://github.com/herd/herdtools7/pull/534#issuecomment-1457876964

[asl-herd] Documentation && explicit module types

[herd] Specify module type and remove dead code

[herd] Introduce order combinator

[asl+herd] Refacto AArch64ASLSem

Failed to extract execution translation into another module.

[asl+herd] Increase space of AArch64+ASL

This commit is a squash of various commits that brings the following
features:
 - Handle correctly DATA ports
 - Support some logical op3 instructions (ORR, EOR, AND, ANDS) ...
 - Support CSET, CINV, CNEG
 - Support LDR/STR register & immediates
 - Correctly supports PSTATE.[N,Z,C,V]

Notes:
 - Failed to extract execution translation into another module
 - Some dark magic in AArch64ASLSem: I use symbolic constants to store
   values that cannot be translated from the AArch64 execution to the
   ASL one. They have a special tag, and their name is a key into a
   hash-table that allow translation in the other direction when
   translating the ASL execution into an AArch64 monad. As long as 1) no
   operation is done on the content of those values and 2) no ASL
   operation has a side effect on those values, it should be fine.
 - We now pass every test in herd/tests/instruction/AArch64 except L043
   and L045 that depends on some unsupported order dependency in SWP.

[asl+herd] Use frozen values inside vectors