[herd,AArch64] Normalise immediate branches and adr instructions to branches with offsets
Doing so we get closer to actual machine code. As a result comparing
instructions for equality is now reliable.

By contrast the ADR instruction returns a (symbolic) label.

This PR also includes a fix of rf enumeration in `-variant self mode`:

Solve explicit code writes and reads first

In self mode, rf pairing of loads and stores is performed by a
two step process:
  1. Pair explicit code accesses. This includes attempt to solve
     the resulting equations
  2. Pair memory accesses, for efficiency loads solved at step 1
     are not considered again.

Doing so, we benefit from not attempting some absurd rf pairing at step 2, when
solving equations at step 1 fails. As a consequence, some non-valid
accesses (typically access to the null location) are discarded early,
which is not the case if all memory accesses are handled together.